Finish the lasagna recipe, unfortunately
This recipe is a war-crime. I think. Like...I might actually try
it, in the same way you have to try hamburger-helper every few years
or so to get the reminder of how bad those childhood broke meals were,
and how privileged you are now that you can order all your favorite
pieces of nigiri in the original Japanese. Hai.